Generalization to multiple backend representations.
Experimental. There are now two ways to interpret a program: 1) as a
direct sampler in the IO monad, and 2) as a KY tree which yields an IO
sampler as well but can be used for other things such as exact
inference. See IORepr.hs and TreeRepr.hs.

Direct IO sampling appears to be slower than the tree route, which is
probably due to some inefficiency in the IO computations being
constructed in the interpretation function. The trees may provide
faster samplers in the long run but I wouldn't expect to see that here
because no reduction is being performed.

The 'Repr' class is at the bottom of Lang.hs.