Add simple exact inference method for boolean-valued trees.
LinEq.hs contains a little solver for sets of linear recurrence
equations that are derived from the structure of a tree. I think the
running time is O(n²) in the size of the tree.

Another easy approach is to pass the equations to Z3, which I may try
and compare performance against the custom solver.